requirements.txt
snowflake-connector-python>=3.0.0
snowflake-snowpark-python>=1.0.0
streamlit>=1.22.0
toml>=0.10.0
pandas>=1.5.0
scikit-learn>=1.0.0
joblib>=1.2.0
kafka-python>=2.0.0
watchdog
